pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "tailo-cli"
version = "0.1.0"
description = "Convert Hanzi or POJ tone-number text into Tâi-lô (台羅)."
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [{name = "OpenAI Codex CLI"}]
dependencies = ["opencc-python-reimplemented>=0.1.7"]

[project.scripts]
tailo = "tailo_cli.__main__:main"

[tool.setuptools]
packages = ["tailo_cli"]
